requirements.txt
selenium==4.15.2
python-docx==0.8.11